+++ .github/workflows/test.yml
@@ -1,0 +1,14 @@
+name: Test
+on: [pull_request]
+
+jobs:
+  integ:
+    name: Integration Tests
+    runs-on: ubuntu-latest
+    container:
+      image: jsii/superchain
+    steps:
+      - uses: actions/checkout@v2
+      - name: Run Tests
+        run: |
+          ./integ-tests/synth.sh
