+++ integ-tests/cdk.json
@@ -1,0 +1,7 @@
+{
+  "app": "python3 integ.py",
+  "context": {
+    "@aws-cdk/core:enableStackNameDuplicates": "true",
+    "aws-cdk:enableDiffNoFail": "true"
+  }
+}
